+++ app/src/main/java/com/example/sportik/domain/interactor/NewsInteractor.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.flow.Flow
 
 interface NewsInteractor {
     suspend fun getNews(): Flow<Resource<List<News>>>
-    suspend fun getNewsWithDetails(id: Long): Flow<Resource<NewsWithContent>>
+    suspend fun getNewsWithDetails(id: Int): Flow<Resource<NewsWithContent>>
 }
