+++ app/src/main/java/com/example/sportik/domain/repository/NewsRepository.kt
@@ -8,5 +8,5 @@ import kotlinx.coroutines.flow.Flow
 interface NewsRepository {
     suspend fun getNews(): Flow<Resource<List<News>>>
 
-    suspend fun getNewsWithDetails(id: Long): Flow<Resource<NewsWithContent>>
+    suspend fun getNewsWithDetails(id: Int): Flow<Resource<NewsWithContent>>
 }
